+++ astroquery/esa/hubble/tests/test_esa_hubble.py
@@ -42,13 +42,29 @@ def get_mockreturn(method, request, url, params, *args, **kwargs):
 def ehst_request(request):
     try:
         mp = request.getfixturevalue("monkeypatch")
-    except AttributeError:  # pytest < 3
+    except AttributeError:
         mp = request.getfuncargvalue("monkeypatch")
     mp.setattr(ESAHubbleClass, '_request', get_mockreturn)
     return mp
 
 
-class TestESAHubble():
+def get_cone_mockreturn(method, request, url, params, *args, **kwargs):
+    file = data_path('cone_search_m31_5.vot')
+    if 'OBSERVATION_ID' in params:
+        file = params['OBSERVATION_ID'] + ".vot"
+    response = data_path(file)
+    shutil.copy(response + '.test', response)
+    return response
+
+
+@pytest.fixture(autouse=True)
+def ehst_cone_search(request):
+    mp = request.getfixturevalue("monkeypatch")
+    mp.setattr(ESAHubbleClass, 'cone_search', get_cone_mockreturn)
+    return mp
+
+
+class TestESAHubble:
 
     def get_dummy_tap_handler(self):
         parameterst = {'query': "select top 10 * from hsc_v2.hubble_sc2",
@@ -58,16 +74,47 @@ class TestESAHubble():
         dummyTapHandler = DummyHubbleTapHandler("launch_job", parameterst)
         return dummyTapHandler
 
-    def test_download_product(self):
+    def test_download_product_errors(self):
+        ehst = ESAHubbleClass(self.get_dummy_tap_handler())
+
+        with pytest.raises(ValueError) as err:
+            ehst.download_product(observation_id="J6FL25S4Q",
+                                  product_type="SCIENCE")
+        assert "This product_type is not allowed" in err.value.args[0]
+
+    def test_download_product_by_calibration(self):
         parameters = {'observation_id': "J6FL25S4Q",
-                      'calibration_level': "RAW",
+                      'cal_level': "RAW",
                       'filename': "J6FL25S4Q.vot",
                       'verbose': True}
         ehst = ESAHubbleClass(self.get_dummy_tap_handler())
-        ehst.download_product(parameters['observation_id'],
-                              parameters['calibration_level'],
-                              parameters['filename'],
-                              parameters['verbose'])
+        ehst.download_product(observation_id=parameters['observation_id'],
+                              calibration_level=parameters['cal_level'],
+                              filename=parameters['filename'],
+                              verbose=parameters['verbose'])
+
+    def test_download_product_by_product_type(self):
+        parameters = {'observation_id': "J6FL25S4Q",
+                      'product_type': "SCIENCE_PRODUCT",
+                      'filename': "J6FL25S4Q.vot",
+                      'verbose': True}
+        ehst = ESAHubbleClass(self.get_dummy_tap_handler())
+        ehst.download_product(observation_id=parameters['observation_id'],
+                              product_type=parameters['product_type'],
+                              filename=parameters['filename'],
+                              verbose=parameters['verbose'])
+        parameters['product_type'] = "PRODUCT"
+        ehst = ESAHubbleClass(self.get_dummy_tap_handler())
+        ehst.download_product(observation_id=parameters['observation_id'],
+                              product_type=parameters['product_type'],
+                              filename=parameters['filename'],
+                              verbose=parameters['verbose'])
+        parameters['product_type'] = "POSTCARD"
+        ehst = ESAHubbleClass(self.get_dummy_tap_handler())
+        ehst.download_product(observation_id=parameters['observation_id'],
+                              product_type=parameters['product_type'],
+                              filename=parameters['filename'],
+                              verbose=parameters['verbose'])
 
     def test_get_postcard(self):
         ehst = ESAHubbleClass(self.get_dummy_tap_handler())
@@ -82,23 +129,15 @@ class TestESAHubble():
         ehst.query_target(name=parameters['name'],
                           verbose=parameters['verbose'])
 
-    @pytest.mark.remote_data
     def test_cone_search(self):
-        coords = coordinates.SkyCoord("00h42m44.51s +41d16m08.45s", frame='icrs')
-
-        parameterst = {'query': "select top 10 * from hsc_v2.hubble_sc2",
-                       'output_file': "test2.vot",
-                       'output_format': "votable",
-                       'verbose': False}
-        dummyTapHandler = DummyHubbleTapHandler("launch_job", parameterst)
-
+        coords = coordinates.SkyCoord("00h42m44.51s +41d16m08.45s",
+                                      frame='icrs')
+        ehst = ESAHubbleClass(self.get_dummy_tap_handler())
         parameters = {'coordinates': coords,
                       'radius': 0.0,
-                      'file_name': 'file_cone',
+                      'filename': 'file_cone',
                       'output_format': 'votable',
                       'cache': True}
-
-        ehst = ESAHubbleClass(dummyTapHandler)
         target_file = data_path('cone_search.vot')
         with open(target_file, mode='rb') as file:
             target_obj = file.read()
@@ -107,40 +146,39 @@ class TestESAHubble():
             ehst._request = MagicMock(return_value=response)
             ehst.cone_search(parameters['coordinates'],
                              parameters['radius'],
-                             parameters['file_name'],
+                             parameters['filename'],
                              parameters['output_format'],
                              parameters['cache'])
+            dummyTapHandler = DummyHubbleTapHandler("cone_search", parameters)
 
-    @pytest.mark.remote_data
     def test_cone_search_coords(self):
         coords = "00h42m44.51s +41d16m08.45s"
 
         parameterst = {'query': "select top 10 * from hsc_v2.hubble_sc2",
                        'output_file': "test2.vot",
                        'output_format': "votable",
-                       'verbose': False}
+                       'verbose': True}
         dummyTapHandler = DummyHubbleTapHandler("launch_job", parameterst)
 
         parameters = {'coordinates': coords,
                       'radius': 0.0,
-                      'file_name': 'file_cone',
+                      'filename': 'file_cone',
+                      'async_job': False,
                       'output_format': 'votable',
-                      'cache': True}
+                      'cache': True,
+                      'verbose': True}
 
         ehst = ESAHubbleClass(dummyTapHandler)
         ehst.cone_search(parameters['coordinates'],
                          parameters['radius'],
-                         parameters['file_name'],
+                         parameters['filename'],
                          parameters['output_format'],
-                         parameters['cache'])
-        parameters[coordinates] = 1234
+                         parameters['async_job'],
+                         parameters['cache'],
+                         parameters['verbose'])
         with pytest.raises(ValueError) as err:
-            ehst.cone_search(parameters['coordinates'],
-                             parameters['radius'],
-                             parameters['file_name'],
-                             parameters['output_format'],
-                             parameters['cache'])
-        assert "coordinate must be either a string or "\
+            ehst._getCoordInput(1234)
+        assert "Coordinates must be either a string or "\
                "astropy.coordinates" in err.value.args[0]
 
     def test_query_hst_tap(self):
@@ -172,6 +210,10 @@ class TestESAHubble():
         dummyTapHandler = DummyHubbleTapHandler("get_tables", parameters2)
         ehst = ESAHubbleClass(self.get_dummy_tap_handler())
         ehst.get_tables(True, True)
+
+    def test_get_artifact(self):
+        ehst = ESAHubbleClass(self.get_dummy_tap_handler())
+        ehst.get_artifact("w0ji0v01t_c2f.fits.gz")
 
     def test_get_columns(self):
         parameters = {'query': "select top 10 * from hsc_v2.hubble_sc2",
@@ -217,9 +259,12 @@ class TestESAHubble():
                        'output_format': "votable",
                        'verbose': False}
         parameters3 = {'query': "select o.*, p.calibration_level, "
-                                "p.data_product_type from ehst.observation "
-                                "AS o LEFT JOIN ehst.plane as p on "
-                                "o.observation_uuid=p.observation_uuid where("
+                                "p.data_product_type, pos.ra, pos.dec "
+                                "from ehst.observation "
+                                "AS o JOIN ehst.plane as p on "
+                                "o.observation_uuid=p.observation_uuid "
+                                "JOIN ehst.position as pos on "
+                                "p.plane_id = pos.plane_id where("
                                 "p.calibration_level LIKE '%PRODUCT%' AND "
                                 "p.data_product_type LIKE '%image%' AND "
                                 "o.intent LIKE '%SCIENCE%' AND (o.collection "
@@ -261,9 +306,12 @@ class TestESAHubble():
                        'output_format': "votable",
                        'verbose': False}
         parameters3 = {'query': "select o.*, p.calibration_level, "
-                                "p.data_product_type from ehst.observation "
-                                "AS o LEFT JOIN ehst.plane as p on "
-                                "o.observation_uuid=p.observation_uuid where("
+                                "p.data_product_type, pos.ra, pos.dec"
+                                " from ehst.observation "
+                                "AS o JOIN ehst.plane as p on "
+                                "o.observation_uuid=p.observation_uuid "
+                                "JOIN ehst.position as pos on p.plane_id "
+                                "= pos.plane_id where("
                                 "p.calibration_level LIKE '%RAW%' AND "
                                 "p.data_product_type LIKE '%image%' AND "
                                 "o.intent LIKE '%SCIENCE%' AND (o.collection "
@@ -289,6 +337,90 @@ class TestESAHubble():
                                 parameters1['verbose'],
                                 parameters1['get_query'])
         assert "Calibration level must be between 0 and 3" in err.value.args[0]
+
+    def test_cone_search_criteria(self):
+        parameters1 = {'target': "m31",
+                       'radius': 7,
+                       'data_product_type': "image",
+                       'obs_collection': ['HST'],
+                       'instrument_name': ['ACS/WFC'],
+                       'filters': ['F435W'],
+                       'async_job': False,
+                       'filename': "output_test_query_by_criteria.vot.gz",
+                       'output_format': "votable",
+                       'verbose': True}
+        test_query = "select o.*, p.calibration_level, p.data_product_type, "\
+                     "pos.ra, pos.dec from ehst.observation AS o JOIN "\
+                     "ehst.plane as p on o.observation_uuid=p.observation_"\
+                     "uuid JOIN ehst.position as pos on p.plane_id = "\
+                     "pos.plane_id where((o.collection LIKE '%HST%') AND "\
+                     "(o.instrument_name LIKE '%WFPC2%') AND "\
+                     "(o.instrument_configuration LIKE '%F606W%') AND "\
+                     "1=CONTAINS(POINT('ICRS', pos.ra, pos.dec),"\
+                     "CIRCLE('ICRS', 10.6847083, 41.26875, "\
+                     "0.11666666666666667)))"
+        parameters3 = {'query': test_query,
+                       'output_file': "output_test_query_by_criteria.vot.gz",
+                       'output_format': "votable",
+                       'verbose': False}
+        ehst = ESAHubbleClass(self.get_dummy_tap_handler())
+        query_criteria_query = "select o.*, p.calibration_level, "\
+                               "p.data_product_type, pos.ra, pos.dec from "\
+                               "ehst.observation AS o JOIN ehst.plane as p "\
+                               "on o.observation_uuid=p.observation_uuid "\
+                               "JOIN ehst.position as pos on p.plane_id = "\
+                               "pos.plane_id where((o.collection LIKE "\
+                               "'%HST%') AND (o.instrument_name LIKE "\
+                               "'%WFPC2%') AND (o.instrument_configuration "\
+                               "LIKE '%F606W%'))"
+        ehst.query_criteria = MagicMock(return_value=query_criteria_query)
+        target = {'RA_DEGREES': '10.6847083', 'DEC_DEGREES': '41.26875'}
+        ehst._query_tap_target = MagicMock(return_value=target)
+        ehst.cone_search_criteria(target=parameters1['target'],
+                                  radius=parameters1['radius'],
+                                  data_product_type=parameters1
+                                  ['data_product_type'],
+                                  obs_collection=parameters1['obs_collection'],
+                                  instrument_name=parameters1
+                                  ['instrument_name'],
+                                  filters=parameters1['filters'],
+                                  async_job=parameters1['async_job'],
+                                  filename=parameters1['filename'],
+                                  output_format=parameters1['output_format'],
+                                  verbose=parameters1['verbose'])
+        dummy_tap_handler = DummyHubbleTapHandler("launch_job", parameters3)
+        dummy_tap_handler.check_call("launch_job", parameters3)
+        c = coordinates.SkyCoord("00h42m44.51s +41d16m08.45s", frame='icrs')
+        ehst.cone_search_criteria(coordinates=c,
+                                  radius=parameters1['radius'],
+                                  data_product_type=parameters1
+                                  ['data_product_type'],
+                                  obs_collection=parameters1['obs_collection'],
+                                  instrument_name=parameters1
+                                  ['instrument_name'],
+                                  filters=parameters1['filters'],
+                                  async_job=parameters1['async_job'],
+                                  filename=parameters1['filename'],
+                                  output_format=parameters1['output_format'],
+                                  verbose=parameters1['verbose'])
+        with pytest.raises(TypeError) as err:
+            ehst.cone_search_criteria(target=parameters1['target'],
+                                      coordinates=123,
+                                      radius=parameters1['radius'],
+                                      data_product_type=parameters1
+                                      ['data_product_type'],
+                                      obs_collection=parameters1
+                                      ['obs_collection'],
+                                      instrument_name=parameters1
+                                      ['instrument_name'],
+                                      filters=parameters1['filters'],
+                                      async_job=parameters1['async_job'],
+                                      filename=parameters1['filename'],
+                                      output_format=parameters1
+                                      ['output_format'],
+                                      verbose=parameters1['verbose'])
+        assert "Please use only target or coordinates as"\
+            "parameter." in err.value.args[0]
 
     def test_query_criteria_no_params(self):
         ehst = ESAHubbleClass(self.get_dummy_tap_handler())